Conflicts:
	src/java/controladores/ClasificadoControlador.java
	src/java/persistencia/daos/ArticuloDAO.java
	src/java/persistencia/daos/MultimediaDAO.java

+++ src/java/persistencia/entidades/Multimedia.java
@@ -14,7 +14,7 @@ public class Multimedia  implements Serializable{
     public Multimedia() {
     }
 
-    public Multimedia(int codigo) {
+    public Multimedia(long codigo) {
         this.codigo = codigo;
     }
